feat: updated scaffold composable; moved scaffold util to new file
